+++ backend/src/main/java/com/colmena/belleza/dto/ProductoDTO.java
@@ -3,12 +3,12 @@ package com.colmena.belleza.dto;
 import java.math.BigDecimal;
 
 public class ProductoDTO {
-    public Long id;
+    public Integer id;
     public String nombre;
     public String descripcion;
     public BigDecimal precio;
     public Integer stock;
     public String codigoBarras;
-    public Long idCategoria;
-    public Long idProveedor;
+    public Integer idCategoria;
+    public Integer idProveedor;
 }
