+++ backend/src/main/java/com/colmena/belleza/repository/CategoriaRepository.java
@@ -3,4 +3,4 @@ package com.colmena.belleza.repository;
 import com.colmena.belleza.entity.Categoria;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface CategoriaRepository extends JpaRepository<Categoria, Long> {}
+public interface CategoriaRepository extends JpaRepository<Categoria, Integer> {}
