+++ backend/src/main/java/com/colmena/belleza/repository/ClienteRepository.java
@@ -3,4 +3,4 @@ package com.colmena.belleza.repository;
 import com.colmena.belleza.entity.Cliente;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ClienteRepository extends JpaRepository<Cliente, Long> {}
+public interface ClienteRepository extends JpaRepository<Cliente, Integer> {}
