+++ backend/src/main/java/com/colmena/belleza/repository/GastoRepository.java
@@ -3,4 +3,4 @@ package com.colmena.belleza.repository;
 import com.colmena.belleza.entity.Gasto;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface GastoRepository extends JpaRepository<Gasto, Long> {}
+public interface GastoRepository extends JpaRepository<Gasto, Integer> {}
