+++ backend/src/main/java/com/colmena/belleza/repository/InventarioMovimientoRepository.java
@@ -3,4 +3,4 @@ package com.colmena.belleza.repository;
 import com.colmena.belleza.entity.InventarioMovimiento;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface InventarioMovimientoRepository extends JpaRepository<InventarioMovimiento, Long> {}
+public interface InventarioMovimientoRepository extends JpaRepository<InventarioMovimiento, Integer> {}
